pyproject.toml
[project]
name = "na-tools"
version = "0.1.0"
description = "Nekro Agent 跨平台自动部署 CLI 工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["click>=8.1", "rich>=13.0", "httpx>=0.27", "pyyaml>=6.0"]

[project.scripts]
na-tools = "na_tools.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/na_tools"]
